Introducing OpenvSwitch Firewall Driver
Until Mitaka, if you want to use OpenvSwitch and Security Groups, Neutron
had to use Linux Bridges too! A complex and slow approach... Not anymore!

So, we're replacing the old and slow OVSHybridIptablesFirewallDriver, by the new
"openvswitch" firewall_driver.

Now, the Security Groups are 100% based on OpenvSwitch features! Not iptables.